Update annotator canvas & enable drawing box tool; Update styles
