+++ ui/src/features/annotator/annotations/annotations.component.tsx
@@ -1,8 +1,6 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { CSSProperties, MouseEvent } from 'react';
-
 import { useAnnotator } from '../annotator-provider.component';
 import { useSelectedAnnotations } from '../select-annotation-provider.component';
 import { Annotation } from './annotation.component';
@@ -14,20 +12,9 @@ type AnnotationsProps = {
     isFocussed: boolean;
 };
 
-const DEFAULT_ANNOTATION_STYLES = {
-    fillOpacity: 0.4,
-    fill: 'var(--annotation-fill)',
-    stroke: 'var(--annotation-stroke)',
-    strokeLinecap: 'round',
-    strokeWidth: 'calc(1px / var(--zoom-scale))',
-    strokeDashoffset: 0,
-    strokeDasharray: 0,
-    strokeOpacity: 'var(--annotation-border-opacity, 1)',
-} satisfies CSSProperties;
-
 export const Annotations = ({ width, height, isFocussed }: AnnotationsProps) => {
     const { annotations } = useAnnotator();
-    const { setSelectedAnnotations, selectedAnnotations } = useSelectedAnnotations();
+    const { selectedAnnotations } = useSelectedAnnotations();
 
     // Order annotations by selection. Selected annotation should always be on top.
     const orderedAnnotations = [
@@ -35,19 +22,11 @@ export const Annotations = ({ width, height, isFocussed }: AnnotationsProps) => 
         ...annotations.filter((a) => selectedAnnotations.has(a.id)),
     ];
 
-    const handleClickOutside = (e: MouseEvent<SVGSVGElement>): void => {
-        if (e.target === e.currentTarget) {
-            setSelectedAnnotations(new Set());
-        }
-    };
-
     return (
-        <svg width={width} height={height} style={DEFAULT_ANNOTATION_STYLES} onClick={handleClickOutside}>
-            <MaskAnnotations annotations={orderedAnnotations} width={width} height={height} isEnabled={isFocussed}>
-                {orderedAnnotations.map((annotation) => (
-                    <Annotation annotation={annotation} key={annotation.id} />
-                ))}
-            </MaskAnnotations>
-        </svg>
+        <MaskAnnotations annotations={orderedAnnotations} width={width} height={height} isEnabled={isFocussed}>
+            {orderedAnnotations.map((annotation) => (
+                <Annotation annotation={annotation} key={annotation.id} />
+            ))}
+        </MaskAnnotations>
     );
 };
