+++ ui/src/features/dataset/media-preview/media-preview.component.tsx
@@ -7,6 +7,7 @@ import { Button, ButtonGroup, Content, Dialog, Divider, Grid, Heading, ToggleBut
 
 import { AnnotatorCanvas } from '../../annotator/annotator-canvas';
 import { AnnotatorProvider } from '../../annotator/annotator-provider.component';
+import { SelectAnnotationProvider } from '../../annotator/select-annotation-provider.component';
 import { ToolSelectionBar } from '../../annotator/tools/tool-selection-bar.component';
 import { response } from '../mock-response';
 
@@ -38,7 +39,9 @@ export const MediaPreview = ({ mediaItem, close }: { mediaItem: Item; close: () 
                             </View>
 
                             <View gridArea={'canvas'}>
-                                <AnnotatorCanvas mediaItem={mediaItem} isFocussed={isFocussed} />
+                                <SelectAnnotationProvider>
+                                    <AnnotatorCanvas mediaItem={mediaItem} isFocussed={isFocussed} />
+                                </SelectAnnotationProvider>
                             </View>
 
                             <View gridArea={'aside'}>
